Moved WaterCalculator to Shared instead because needed in both client and server
