+++ src/main/java/se/myhappyplants/client/view/LibraryPlantPane.java
@@ -13,7 +13,7 @@ import javafx.util.Duration;
 import se.myhappyplants.client.controller.MyPlantsTabPaneController;
 import se.myhappyplants.client.model.BoxTitle;
 import se.myhappyplants.client.model.PictureRandomizer;
-import se.myhappyplants.client.model.WaterCalculator;
+import se.myhappyplants.shared.WaterCalculator;
 import se.myhappyplants.shared.Plant;
 import se.myhappyplants.shared.PlantDetails;
