+++ src/main/java/se/myhappyplants/server/services/PlantRepository.java
@@ -1,13 +1,11 @@
 package se.myhappyplants.server.services;
 
-import se.myhappyplants.client.view.LightTextFormatter;
-import se.myhappyplants.client.model.WaterCalculator;
+import se.myhappyplants.shared.WaterCalculator;
 import se.myhappyplants.shared.Plant;
 import se.myhappyplants.shared.PlantDetails;
 
 import java.io.IOException;
 import java.net.UnknownHostException;
-import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -82,20 +80,20 @@ public class PlantRepository {
         }
     }
 
-    public int getWaterFrequency(String plantId) throws IOException, InterruptedException {
-        int water = -1;
+    public long getWaterFrequency(String plantId) throws IOException, InterruptedException {
+        long waterFrequency = -1;
         String query = "SELECT water_frequency FROM species WHERE id = '" + plantId + "';";
         try {
             ResultSet resultSet = database.executeQuery(query);
             while (resultSet.next()) {
                 String waterText = resultSet.getString("water_frequency");
-
-                water = (isNumeric(waterText)) ? Integer.parseInt(waterText) : -1;
+                int water = (isNumeric(waterText)) ? Integer.parseInt(waterText) : -1;
+                waterFrequency = WaterCalculator.calculateWaterFrequencyForWatering(water);
             }
         }
         catch (SQLException throwables) {
             throwables.printStackTrace();
         }
-        return water;
+        return waterFrequency;
     }
 }
